requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
numpy>=1.26,<3.0
polars>=1.0,<2.0
pydantic>=2.0,<3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
